requirements.txt
discord.py>=2.5.2
python-dotenv>=1.0.0
aiomysql>=0.1.1
asyncio>=3.4.3
pytz>=2021.3
psutil>=5.9.0
aiohttp>=3.8.0
mysql-connector-python>=8.0.28
PyNaCl>=1.5.0 